// src/interfaces.ts
export interface QuizResponse {
  response_code: number;
  results:       Result[];
}

export interface Result {
  category:          string;
  type:              Type;
  difficulty:        Difficulty;
  question:          string;
  correct_answer:    string;
  incorrect_answers: string[];
}

export enum Difficulty {
  Easy = "easy",
  Hard = "hard",
  Medium = "medium",
}

export enum Type {
  Multiple = "multiple",
}